Core: add a basic stub of a core library for bool
This adds a basic library skeleton with a boolean type defined
sufficiently for use in a small context.  It allows for equality
comparisons and the use of bool in a conditional expression.

The runtime is needed for the standard library.  Provide symbolic stubs
for the runtime for the known metadata.  This is sufficient for linking
the DSO, but not run, as that will require the definitions to be fleshed
out.