Lower required cmake version to 3.16.
Some of the Linux Swift project build bots only have version 3.16
installed. It's unclear if 3.18 is a requirement for Linux too (the docs
just say macOS), but either way, they should probably updated. That
being said, this is an easier than updating all the build bots (or,
more realistically, asking Mishal to).