build: correct the linking of swiftCore
This fixes the configuration of the build to build the swiftCore shared library
properly.  We should treat the library as being a C (or C++) library.  This
ensures that we use the C/C++ linker driver to link the library rather than the
Swift linker driver.  This will avoid the incorrect embedding of the Swift
registrar for non-Darwin targets.